requirements.txt
httplib2==0.22.0
aiogram==2.25.1
asyncpg==0.29.0
python-dotenv==1.0.1